V1.5 La posicion del cuadrado se coloca desde el Preview.

+++ Cliente/Models/DibuRec.cs
@@ -11,8 +11,8 @@ namespace Servidor.Models
     {
         public int? Ancho { get; set; } = 50;
         public int? Alto { get; set; } = 50;
-        public int? X { get; set; } = 0;
-        public int? Y { get; set; } = 0;
+        public double? X { get; set; } = 0;
+        public double? Y { get; set; } = 0;
         public string? Color { get; set; } = "#9BADB7";
 
     }
