+++ Servidor/Models/DibuRec.cs
@@ -8,8 +8,8 @@ namespace Servidor.Models
 {
     internal class DibuRec
     {
-        public double? Ancho { get; set; }
-        public double? Alto { get; set; }
+        public int? Ancho { get; set; }
+        public int? Alto { get; set; }
         public double? X { get; set; }
         public double? Y { get; set; }
         public string? Color { get; set; }
